+++ messages/de/yii-dataview.php
@@ -6,4 +6,10 @@ return [
     'No results found.' => 'Keine Ergebnisse gefunden.',
     'Page <b>{currentPage}</b> of <b>{totalPages}</b>' => 'Seite <b>{currentPage}</b> von <b>{totalPages}</b>',
     'Actions' => 'Aktionen',
+    'Pagination' => 'Seitennavigation',
+    'First page' => 'Erste Seite',
+    'Previous page' => 'Vorherige Seite',
+    'Next page' => 'Nächste Seite',
+    'Last page' => 'Letzte Seite',
+    'Page {page}' => 'Seite {page}',
 ];