+++ tests/GridView/GridViewTest.php
@@ -918,13 +918,13 @@ final class GridViewTest extends TestCase
 
         $this->assertStringContainsString(
             <<<HTML
-            <nav>
-            <a aria-disabled="true" href="/route?">⟪</a>
-            <a aria-disabled="true" href="/route?">⟨</a>
-            <a aria-current="page" href="/route?">1</a>
-            <a href="/route?page=2">2</a>
-            <a href="/route?page=2">⟩</a>
-            <a href="/route?page=2">⟫</a>
+            <nav aria-label="Pagination">
+            <a aria-label="First page" aria-disabled="true" href="/route?">⟪</a>
+            <a aria-label="Previous page" aria-disabled="true" href="/route?">⟨</a>
+            <a aria-label="Page 1" aria-current="page" href="/route?">1</a>
+            <a aria-label="Page 2" href="/route?page=2">2</a>
+            <a aria-label="Next page" href="/route?page=2">⟩</a>
+            <a aria-label="Last page" href="/route?page=2">⟫</a>
             </nav>
             HTML,
             $html,